refactor(backend): address CodeRabbit findings on agent-config (#104)
- Log unexpected config-route failures server-side via pino before the
  generic 500 envelope, in both the per-rig and fleet PATCH catch blocks,
  so DB/audit failures are diagnosable without leaking details to clients.
- Split the pure agent-config helpers (validateRawFlags, mergeEffectiveConfig,
  mergeWhitelist, and the merge/parse builders) into agent-config-helpers.ts
  so they no longer transitively import env. Importing the helpers no longer
  triggers env validation at load time (GOTCHAS.md Pattern B), keeping the
  pure unit tests env-independent. The service module re-exports the public
  surface, so route and agent-API importers are unchanged.
